Adding transient storage to StorageProvider
